Ajout du test du pourcentage d'occupation d'Une salle dans la classe salle.

+++ ClAssignateur-Projet9/ClAssignateur-domain/src/main/java/org/ClAssignateur/domain/Salle.java
@@ -16,7 +16,7 @@ public class Salle {
 		return capacite >= nbParticipants;
 	}
 
-	public double pourcentageOccupation(int nbParticipants) {
+	public long pourcentageOccupation(int nbParticipants) {
 		return nbParticipants / capacite;
 	}
